requirements.txt
flet >=0.70.0.dev0
python-dotenv
requests